requirements.txt
click==8.0.3
Flask==2.2.5
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.19.2
Werkzeug==2.0.2